M1

+++ morphing/morph.cpp
@@ -51,7 +51,7 @@ int main(int argc, char** argv){
 		for(int j=1; j<input_1.cols()-1; j++){
 			for(int k=0; k<gaussian.rows(); k++){
 				for(int l=0; l<gaussian.cols(); l++){
-					lowPass_1[i][j] += input_1[i][j] * gaussian[gaussian.rows()-k-1][gaussian.cols()-l-1];
+					lowPass_1[i][j] += input_1[i][j] * (double)gaussian[gaussian.rows()-k-1][gaussian.cols()-l-1];
 				}
 			}	
 		}
@@ -62,7 +62,7 @@ int main(int argc, char** argv){
 		for(int j=1; j<input_2.cols()-1; j++){
 			for(int k=0; k<gaussian.rows(); k++){
 				for(int l=0; l<gaussian.cols(); l++){
-					lowPass_2[i][j] += input_2[i][j] * gaussian[gaussian.rows()-k-1][gaussian.cols()-l-1];
+					lowPass_2[i][j] += input_2[i][j] * (double)gaussian[gaussian.rows()-k-1][gaussian.cols()-l-1];
 				}
 			}	
 		}
